Added Body Validator

+++ api/src/application/controllers/AdminController.ts
@@ -1,10 +1,14 @@
 import { Request, Response } from "express";
 import Admin from "../models/AdminModel"
 import { compare } from "bcrypt";
+import { validateBody } from "../../core/utils";
 
 const getAdmin = async (req: Request, res: Response) => {
     try{
         const { username, password } = req.body;
+        if(!validateBody(req.body, username, password)) {
+            return res.status(408).json({ error: 'Request body incomplete' });
+        }
         const admin = await Admin.findOne({ where: { username } });
         if (!admin) {
             return res.status(404).json({ error: 'Admin not found' });
@@ -22,6 +26,9 @@ const getAdmin = async (req: Request, res: Response) => {
 const createAdmin = async (req: Request, res: Response) => {
     try{
         const { username, password, email } = req.body;
+        if(!validateBody(req.body, username, password, email)) {
+            return res.status(408).json({ error: 'Request body incomplete' });
+        }
         const admin = await Admin.create({ username, password, email });
         return res.status(201).json(admin);
     } catch (error) {
@@ -32,6 +39,9 @@ const createAdmin = async (req: Request, res: Response) => {
 const updateAdmin = async (req: Request, res: Response) => {
     try{
         const { username, password, email } = req.body;
+        if(!validateBody(req.body, username, password, email)) {
+            return res.status(408).json({ error: 'Request body incomplete' });
+        }
         const admin = await Admin.update({ password, email }, { where: { username } });
         return res.status(200).json(admin);
     } catch (error) {
@@ -42,6 +52,9 @@ const updateAdmin = async (req: Request, res: Response) => {
 const deleteAdmin = async (req: Request, res: Response) => {
     try{
         const { username } = req.body;
+        if(!validateBody(req.body, username)) {
+            return res.status(408).json({ error: 'Request body incomplete' });
+        }
         const admin = await Admin.destroy({ where: { username } });
         return res.status(200).json(admin);
     } catch (error) {
