+++ api/src/application/controllers/PlayerController.ts
@@ -1,10 +1,14 @@
 import { Request, Response } from "express";
 import Player from "../models/PlayerModel"
 import { CustomRequest } from "../../core/types";
+import { validateBody } from "../../core/utils";
 
 const getPlayer = async (req: CustomRequest, res: Response) => {
     try{
         const { email } = req.body;
+        if(!validateBody(req.body, email)) {
+            return res.status(408).json({ error: 'Request body incomplete' });
+        }
         if(email === req.email) {
             const player = await Player.findOne({ where: { email } });
             if (!player) {
@@ -21,6 +25,9 @@ const getPlayer = async (req: CustomRequest, res: Response) => {
 const createPlayer = async (req: CustomRequest, res: Response) => {
     try{
         const { username, email, fullname, age, team, favorite_pokemon } = req.body;
+        if(!validateBody(req.body, username, email, fullname, age)) {
+            return res.status(408).json({ error: 'Request body incomplete' });
+        }
         if(username === req.username && email === req.email) {
             const player = await Player.create({ username, email, fullname, age, team, favorite_pokemon });
             return res.status(201).json(player);
@@ -35,6 +42,9 @@ const createPlayer = async (req: CustomRequest, res: Response) => {
 const updatePlayer = async (req: CustomRequest, res: Response) => {
     try{
         const { username, email, fullname, age, team, favorite_pokemon } = req.body;
+        if(!validateBody(req.body, username, email, fullname, age)) {
+            return res.status(408).json({ error: 'Request body incomplete' });
+        }
         if(username === req.username && email === req.email) {
             const player = await Player.update({ fullname, age, team, favorite_pokemon }, { where: { email } });
             if( player == 1) {
@@ -54,6 +64,9 @@ const updatePlayer = async (req: CustomRequest, res: Response) => {
 const deletePlayer = async (req: Request, res: Response) => {
     try{
         const { email } = req.body;
+        if(!validateBody(req.body, email)) {
+            return res.status(408).json({ error: 'Request body incomplete' });
+        }
         const player = await Player.destroy({ where: { email } });
         return res.status(200).json(player);
     } catch (error) {
